chore: move `NotebookDesktopIconsMenu` to notebook_layout
